swml/parsing/inlines-4.dat

#data
aa[FRAC@en[bbb]@fr[de]]ccc
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <math mfrac>
|         xml lang="en"
|         <math mtext>
|           "bbb"
|         <math mtext>
|           xml lang="fr"
|           "de"
|       "ccc"

#data
aa[FRAC@en[bbb]]ccc
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <math mfrac>
|         xml lang="en"
|         <math mtext>
|           "bbb"
|       "ccc"

#data
aa[ROOT@en[bbb]@fr[de]]ccc
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <math mroot>
|         xml lang="en"
|         <math mtext>
|           "bbb"
|         <math mtext>
|           xml lang="fr"
|           "de"
|       "ccc"

#data
aa[ROOT@en[bbb]]ccc
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <math mroot>
|         xml lang="en"
|         <math mtext>
|           "bbb"
|       "ccc"

#data
aa[ROOT@en[bbb]@fr[de][zz]]ccc
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <math mroot>
|         xml lang="en"
|         <math mtext>
|           "bbb"
|         <math mtext>
|           xml lang="fr"
|           "de"
|         <math mtext>
|           "zz"
|       "ccc"

#data
aa[SQRT@en[bbb]@fr[de]]ccc
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <math msqrt>
|         xml lang="en"
|         <math mtext>
|           "bbb"
|         <math mtext>
|           xml lang="fr"
|           "de"
|       "ccc"

#data
aa[SQRT@en[bbb]]ccc
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <math msqrt>
|         xml lang="en"
|         <math mtext>
|           "bbb"
|       "ccc"

#data
aa[SQRT@en[bbb]@fr[de][zz]]ccc
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <math msqrt>
|         xml lang="en"
|         <math mtext>
|           "bbb"
|         <math mtext>
|           xml lang="fr"
|           "de"
|         <math mtext>
|           "zz"
|       "ccc"

#data
aa[UNDEROVER@en[bbb]@fr[de]]ccc
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <math munderover>
|         xml lang="en"
|         <math mtext>
|           "bbb"
|         <math mtext>
|           xml lang="fr"
|           "de"
|       "ccc"

#data
aa[UNDEROVER@en[bbb]]ccc
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <math munderover>
|         xml lang="en"
|         <math mtext>
|           "bbb"
|       "ccc"

#data
aa[UNDEROVER@en[bbb]@fr[de][zz]]ccc
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <math munderover>
|         xml lang="en"
|         <math mtext>
|           "bbb"
|         <math mtext>
|           xml lang="fr"
|           "de"
|         <math mtext>
|           "zz"
|       "ccc"

#data
aa[UNDER@en[bbb]@fr[de]]ccc
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <math munder>
|         xml lang="en"
|         <math mtext>
|           "bbb"
|         <math mtext>
|           xml lang="fr"
|           "de"
|       "ccc"

#data
aa[UNDER@en[bbb]]ccc
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <math munder>
|         xml lang="en"
|         <math mtext>
|           "bbb"
|       "ccc"

#data
aa[UNDER@en[bbb]@fr[de][zz]]ccc
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <math munder>
|         xml lang="en"
|         <math mtext>
|           "bbb"
|         <math mtext>
|           xml lang="fr"
|           "de"
|         <math mtext>
|           "zz"
|       "ccc"
